+++ src/devguide/fulcro_devguide/O09_Forms_Predefined_Fields.cljs
@@ -1,9 +1,9 @@
 (ns fulcro-devguide.O09-Forms-Predefined-Fields
   (:require
     [fulcro.client.dom :as dom]
-    [devcards.core :as dc :refer-macros [defcard defcard-doc]]
-    [fulcro.client.primitives :as prim :refer [defui]]
-    [fulcro.client.cards :refer [fulcro-app]]
+    [devcards.core :as dc :refer-macros [defcard-doc]]
+    [fulcro.client.primitives :as prim :refer [defui defsc]]
+    [fulcro.client.cards :refer [defcard-fulcro]]
     [fulcro.client.core :as fc]
     [fulcro.ui.forms :as f]
     [fulcro.ui.elements :as ele]
@@ -29,8 +29,8 @@
     (dom/div #js {:className "col-sm-10"} (apply f/form-field comp form name params))))
 
 (defui ^:once KitchenSink
-  static fc/InitialAppState
-  (initial-state [this params] (f/build-form this {:db/id 1 :short-story (fc/get-initial-state FileUploadInput {:id :story})}))
+  static prim/InitialAppState
+  (initial-state [this params] (f/build-form this {:db/id 1 :short-story (prim/get-initial-state FileUploadInput {:id :story})}))
   static f/IForm
   (form-spec [this] [(f/id-field :db/id)
                      (f/text-input :text)
@@ -70,8 +70,8 @@
 (def ui-sink (prim/factory KitchenSink {:keyfn :db/id}))
 
 (defui ^:once CommitRoot
-  static fc/InitialAppState
-  (initial-state [this _] {:sink (fc/initial-state KitchenSink {:db/id 1})})
+  static prim/InitialAppState
+  (initial-state [this _] {:sink (prim/get-initial-state KitchenSink {:db/id 1})})
   static prim/IQuery
   (query [this] [:ui/react-key
                  {:sink (prim/get-query KitchenSink)}])
@@ -99,10 +99,10 @@
   "
   (dc/mkdn-pprint-source KitchenSink))
 
-(defcard form-changes
-  (fulcro-app CommitRoot
-    :networking {:remote      (net/make-fulcro-network "/api" :global-error-callback identity)
-                 :file-upload (file-upload-networking)})
+(defcard-fulcro form-changes
+  CommitRoot
   {}
-  {:inspect-data false})
+  {:fulcro       {:networking {:remote      (net/make-fulcro-network "/api" :global-error-callback identity)
+                               :file-upload (file-upload-networking)}}
+   :inspect-data false})