+++ src/test/fulcro/client/impl/application_spec.cljs
@@ -1,7 +1,7 @@
 (ns fulcro.client.impl.application-spec
   (:require
     [fulcro.client.core :as fc]
-    [fulcro.client.primitives :as prim :refer [defui]]
+    [fulcro.client.primitives :as prim :refer [defui defsc]]
     [fulcro-spec.core :refer-macros [specification behavior assertions provided component when-mocking]]
     [fulcro.client.dom :as dom]
     [fulcro.i18n :as i18n]